Convert to Jinja for better whitespace control
