+++ apps/llms_txt/apps.py
@@ -1,7 +1,7 @@
 from django.apps import AppConfig
 
 
-class LlmsTxtConfig(AppConfig):
+class LLMsTxtConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "apps.llms_txt"
     verbose_name = "llms.txt"
