+++ apps/llms_txt/jinja2/llms_txt/page.md.jinja
@@ -1,0 +1,27 @@
+# {{ page.seo_title if page.seo_title else page.title }}
+
+Page URL: {{ page.full_url }}
+{% if page.search_description %}
+> {{ page.search_description }}
+{% endif %}
+------------------------------------------------------------------------
+{% if page.introduction %}
+{{ page.introduction|richtext_markdown }}
+{%- endif %}
+{%- for block in page.body %}
+{%- if block.block_type == 'text' %}
+{{ block.value|richtext_markdown }}
+{%- elif block.block_type == 'alert' %}
+{{block.value.alert_type}}: {{ block.value.alert_body|richtext_markdown }}
+{%- endif %}
+{%- endfor %}
+
+{#- Homepage support -#}
+{%- if page.sections -%}
+{%- for block in page.sections[0].value -%}
+## [{{ block.page.title }}]({{ block.page.full_url }})
+
+{{ block.text }}
+
+{% endfor %}
+{%- endif -%}
