--- apps/llms_txt/templates/llms_txt/llms_full_txt.html
@@ -1,11 +1,0 @@
-{% load wagtailcore_tags wagtailroutablepage_tags %}
-{% wagtail_site as current_site %}
-
-# {{ current_site.root_page.title }}
-
-{% if current_site.root_page.search_description%}> {{ current_site.root_page.search_description }}{% endif %}
-
-{% for page in pages %}
-{{ page.to_markdown }}
----
-{% endfor %}
