--- apps/llms_txt/templates/llms_txt/llms_txt.html
@@ -1,13 +1,0 @@
-{% load wagtailcore_tags wagtailroutablepage_tags %}
-{% wagtail_site as current_site %}
-
-# {{ current_site.root_page.title }}
-
-{% if current_site.root_page.search_description%}> {{ current_site.root_page.search_description }}{% endif %}
-
-View the full site contents: [llms-full.txt]({{ current_site.root_url }}/llms-full.txt)
-
-## Table of Contents
-
-{% for page in pages %}
-- [{% if page.seo_title %}{{ page.seo_title }}{% else %}{{ page.title }}{% endif %}]({{ current_site.root_url }}{% routablepageurl page "markdown" %}){% if page.search_description %}: {{ page.search_description }}{% endif %}{% endfor %}
