--- apps/llms_txt/templates/llms_txt/markdown.html
@@ -1,23 +1,0 @@
-{% load llms_txt_tags wagtailcore_tags %}
-# {% if page.seo_title %}{{ page.seo_title }}{% else %}{{ page.title }}{% endif %}
-
-Page URL: {{ page.full_url }}
-
-{% if page.search_description %}
-> {{ page.search_description }}
-{% endif %}
-
----
-
-{% if page.introduction %}{{ page.introduction|richtext_markdown }}{% endif %}
-
-{% for block in page.body %}
-{% if block.block_type == 'text' %}{{ block.value|richtext_markdown }}
-{% elif block.block_type == 'alert' %}{{ block.value.alert_body|richtext_markdown }}
-{% endif %}{% endfor %}
-
-{% for section in page.sections %}
-{% if section.page %}## {{ section.page.title }}{% endif %}
-
-{{ section.text|richtext_markdown }}
-{% endfor %}
